CCNN/requirements.txt
torch
numpy
scipy
matplotlib
Pillow
scikit-learn
# Note: UHD/gnuradio and USRP drivers must be installed via OS packages (not pip)